requirements.txt
PySide6
discord.py
websocket-client